scripts/ai_news_aggregator/requirements.txt
requests==2.31.0
feedparser==6.0.11
google-generativeai==0.5.4
python-telegram-bot==21.3